% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_nfl_players.R
\name{get_all_nfl_players}
\alias{get_all_nfl_players}
\title{Gather NFL Player Data}
\usage{
get_all_nfl_players()
}
\value{
Returns a list containing information about the NFL players.
}
\description{
Gather all information concerning NFL players. Per Sleeper, this be called once per day at most given
the amount of data that it returns (5MB).
}
\examples{
get_all_nfl_players()

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2021
}
\keyword{nfl}
\keyword{players}
