% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draft.R
\name{get_draft}
\alias{get_draft}
\title{Gather Draft Information}
\usage{
get_draft(draft_id)
}
\arguments{
\item{draft_id}{Draft ID generated by Sleeper (integer)}
}
\value{
Returns a list containing information about the draft.
}
\description{
Given a draft ID, gather draft information for selection.
}
\examples{
\dontrun{get_draft(688281872463106048)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2021
}
\keyword{draft}
