% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draft_pick_trades.R
\name{get_draft_pick_trades}
\alias{get_draft_pick_trades}
\title{Gather Draft Pick Trades}
\usage{
get_draft_pick_trades(draft_id)
}
\arguments{
\item{draft_id}{Draft ID generated by Sleeper (numeric or character)}
}
\value{
Returns a data frame containing information about the draft pick trade(s).
}
\description{
Given a draft ID, gather draft pick trades for that draft.
}
\examples{
\dontrun{get_draft_pick_trades(688281872463106048)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2021
}
\keyword{draft}
\keyword{pick}
\keyword{trades}
