% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draft_picks.R
\name{get_draft_picks}
\alias{get_draft_picks}
\title{Gather Draft Picks}
\usage{
get_draft_picks(draft_id)
}
\arguments{
\item{draft_id}{Draft ID generated by Sleeper (numeric or character)}
}
\value{
Returns a data frame containing information about the draft picks.
}
\description{
Given a draft ID, gather draft picks (along with other metadata concerning the picks).
}
\examples{
\dontrun{get_draft_picks(688281872463106048)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2021
}
\keyword{draft}
\keyword{picks}
