% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_league_users_data.R
\name{get_league_users_data}
\alias{get_league_users_data}
\title{Gather User Data for League}
\usage{
get_league_users_data(league_id)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (integer)}
}
\value{
Returns a data frame containing information about the users in the league.
}
\description{
Given a league ID, grab the data concerning each user. This includes information about the users' mascot,
display name, avatar, if they are an owner, if they are a bot, and more.
}
\examples{
get_league_users_data(688281863499907072)

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{league}
\keyword{users}
