% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_main_data.R
\name{get_main_data}
\alias{get_main_data}
\title{Generate Main Data}
\usage{
get_main_data(league_id)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (numeric).}
}
\value{
Returns a data frame containing roster, league user,
        and overall league information.
}
\description{
Given the league ID, generate a main data set that can be used for
generating quick insights. This includes teams, fantasy points,
avatars, divisions, owner IDs, and more.
}
\examples{
\dontrun{get_main_data(688281863499907072)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2023
}
\keyword{full}
\keyword{league}
\keyword{main}
\keyword{roster}
\keyword{user}
