% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matchup_data.R
\name{get_matchup_data}
\alias{get_matchup_data}
\title{Gather Matchup Data in a League for a Week}
\usage{
get_matchup_data(league_id, week)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (integer)}

\item{week}{Matchup week (integer)}
}
\value{
Returns a data frame containing information about the matchup.
}
\description{
Given a league ID and week number, grab the matchup data, including player points, total points,
starting rosters, and more.
}
\examples{
get_matchup_data(688281863499907072, 2)
get_matchup_data(688281863499907072, 1)

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{league}
\keyword{matchup}
