% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roster_data.R
\name{get_roster_data}
\alias{get_roster_data}
\title{Gather Roster Data for League}
\usage{
get_roster_data(league_id)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (integer)}
}
\value{
Returns a data frame containing information about the rosters in the league.
}
\description{
Given a league ID, grab the data concerning each roster in that league. This includes information with the 
roster wins, losses, ties, fantasy points, and more.
}
\examples{
get_roster_data(688281863499907072)

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{league}
\keyword{rosters}
\keyword{users}
