% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{get_rosters}
\alias{get_rosters}
\title{Gather Roster Data for League}
\usage{
get_rosters(league_id)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (numeric or character)}
}
\value{
Returns a data frame containing information about the rosters in the league.
}
\description{
Given a league ID, grab the data concerning each roster in that league. This includes information with the 
roster wins, losses, ties, fantasy points, and more.
}
\examples{
\dontrun{get_rosters(688281863499907072)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{league}
\keyword{rosters}
\keyword{users}
