% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traded_picks.R
\name{get_traded_picks}
\alias{get_traded_picks}
\title{Gather League Trades}
\usage{
get_traded_picks(league_id)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (numeric or character)}
}
\value{
Returns a list containing trades.
}
\description{
Given a league ID, grab the trades that have occured.
}
\examples{
\dontrun{get_traded_picks(688281863499907072)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{league}
\keyword{trades}
