% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transaction_data.R
\name{get_transaction_data}
\alias{get_transaction_data}
\title{Gather League Transactions for Specific Round}
\usage{
get_transaction_data(league_id, round)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (integer)}

\item{round}{round of transactions to gather (can also be thought of as week depending on when your league processes transactions) (integer)}
}
\value{
Returns a data frame containing information about the round of transactions for that league.
}
\description{
Given a league ID and round, grab the transaction data concerning that particular league.
This includes adds or drops, whether it was a free agent or a waiver, notes, and more.
}
\examples{
get_transaction_data(688281863499907072, 2)

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{league}
\keyword{transactions}
