% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_data.R
\name{get_user_data}
\alias{get_user_data}
\title{Gather User Data}
\usage{
get_user_data(user_id)
}
\arguments{
\item{user_id}{User ID generated by Sleeper (integer)}
}
\value{
Returns a list containing information about the user.
}
\description{
Given a user ID, grab the data concerning that particular user. This includes username, user ID, display 
name, and more.
}
\examples{
get_user_data(688556535013502976)

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{user}
\keyword{users}
