% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpts_against.R
\name{plot_fpts_against}
\alias{plot_fpts_against}
\title{Plot League Fantasy Points Against by User}
\usage{
plot_fpts_against(league_id, title = "", subtitle = "")
}
\arguments{
\item{league_id}{League ID generated by Sleeper (integer)}

\item{title}{Optional title of the plot (string)}

\item{subtitle}{Optional subtitle of the plot (string)}
}
\value{
Returns a bar chart showing fantasy points against by each user.
}
\description{
Given a league ID, plot the number of fantasy points scored against so far by each user. Note that this 
plot was created specifically with a NFL fantasy league, so results may vary when Sleeper decides to 
incorporate other fantasy sports data.
}
\examples{
plot_fpts_against(688281863499907072, "Fantasy Points Against by User")

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{against}
\keyword{fantasy}
\keyword{plot}
\keyword{points}
\keyword{standings}
