% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generate_master_data.R
\name{plot_generate_master_data}
\alias{plot_generate_master_data}
\title{Generate Master Data Frame for Plotting}
\usage{
plot_generate_master_data(league_id)
}
\arguments{
\item{lookback_hours}{Number of hours to look back (numeric). Default is 24}

\item{limit}{Number of results you would like (numeric). Default is 10}
}
\value{
Returns a data frame containing roster, league user, and overall league information.
}
\description{
Given the league ID, generate a master data set that is used by plotting functions and generating fast
insights.
}
\examples{
\dontrun{plot_generate_master_data(688281863499907072)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2023
}
\keyword{full}
\keyword{league}
\keyword{master}
\keyword{roster}
\keyword{user}
