% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_league_dashboard.R
\name{plot_league_dashboard}
\alias{plot_league_dashboard}
\title{Plot League Dashboard}
\usage{
plot_league_dashboard(league_id)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (numeric).}
}
\value{
Returns a dashboard of the current league state.
}
\description{
Given the league ID, generate a dashboard with various
plots/tables of the current league. Note that since this
is a dashboard, your R session will continuously run until
the console is stopped. Stoppage can be achieved by pressing
Control + C on the keyboard.
}
\examples{
\dontrun{plot_league_dashboard(688281863499907072)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, March 2024
}
\keyword{dashboard}
\keyword{league}
\keyword{plot}
