% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_league_records_table.R,
%   R/plot_user_streaks.R
\name{plot_league_records_table}
\alias{plot_league_records_table}
\title{Plot League Information Table}
\usage{
plot_league_records_table(league_id)

plot_league_records_table(league_id)
}
\arguments{
\item{league_id}{League ID generated from Sleeper (numeric).}

\item{lookback_hours}{Number of hours to look back (numeric). Default is 24}

\item{limit}{Number of results you would like (numeric). Default is 10}
}
\value{
Returns a table containing team names and their associated records.

Returns a plot showing streaks by user.
}
\description{
Given the league ID, generate a table showing team names and their associated statistics (wins, losses, points for/against, etc.)

Given the league ID, generate a plot showing the current user streaks.
}
\examples{
\dontrun{plot_league_records_table(688281863499907072)}

\dontrun{get_trending_players("nfl", "add", 24, 25)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2023

Nick Bultman, \email{njbultman74@gmail.com}, January 2024
}
\keyword{league}
\keyword{records}
\keyword{statistics}
\keyword{streaks}
\keyword{users}
