% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_league_streaks_table.R
\name{plot_league_streaks_table}
\alias{plot_league_streaks_table}
\title{Plot League Streaks Table}
\usage{
plot_league_streaks_table(league_id, font_color = "inherit")
}
\arguments{
\item{league_id}{League ID generated from Sleeper (numeric).}

\item{font_color}{Header font color, hex code or name. Default is "inherit" (string).}
}
\value{
Returns a table containing display names and their streaks.
}
\description{
Given the league ID, generate a table showing display names
and their associated streaks (2W, 5L, etc.). Note that the font
color for the streaks are currently at zero or greater = green
and L = red. The font color argument is used for the headers.
}
\examples{
\dontrun{plot_league_streaks_table(688281863499907072)}
\dontrun{plot_league_streaks_table(688281863499907072, font_color = "red")}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2024
}
\keyword{information}
\keyword{league}
\keyword{statistics}
