% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nfl_player_data_table.R
\name{plot_nfl_player_data_table}
\alias{plot_nfl_player_data_table}
\title{Plot NFL Player Information in Table}
\usage{
plot_nfl_player_data_table(font_color = "inherit")
}
\arguments{
\item{font_color}{Font color, name or hex (string).}
}
\value{
Returns a table showing general NFL player information.
}
\description{
This function will generate a table containing general
information (name, age, college, etc.) concerning all
the NFL players that Sleeper has data on.
}
\examples{
\dontrun{plot_nfl_player_data_table()}
\dontrun{plot_nfl_player_data_table(font_color = "green")}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, March 2024
}
\keyword{general}
\keyword{players}
