% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trending_players.R
\name{plot_trending_players}
\alias{plot_trending_players}
\title{Plot Trending NFL Player Information}
\usage{
plot_trending_players(lookback_hours = 24, limit = 10)
}
\arguments{
\item{lookback_hours}{Number of hours to look back (numeric). Default is 24}

\item{limit}{Number of results you would like (numeric). Default is 10}
}
\value{
Returns a plot containing information (name and count) about the NFL trending players.
}
\description{
Given the number of hours to look back, and a limit of how many players to return, display a
figure showing the top NFL additions and drops.
}
\examples{
\dontrun{get_trending_players("nfl", "add", 24, 25)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2023
}
\keyword{players}
\keyword{trending}
