% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trending_players_table.R
\name{plot_trending_players_table}
\alias{plot_trending_players_table}
\title{Plot Trending NFL Player Information in Table}
\usage{
plot_trending_players_table(
  lookback_hours = 24,
  limit = 10,
  font_color = "black"
)
}
\arguments{
\item{lookback_hours}{Number of hours to look back (numeric). Default is 24}

\item{limit}{Number of results you would like (numeric). Default is 10, max is 50}

\item{font_color}{Font color, name or hex (string). Default is black}
}
\value{
Returns a plot containing information (name and
        count) about the NFL trending players.
}
\description{
Given the number of hours to look back, and a limit of
how many players to return, display a figure showing the
top additions and drops in a table format.
}
\examples{
\dontrun{plot_trending_players_table(lookback_hours = 24, limit = 10)}
\dontrun{plot_trending_players_table(lookback_hours = 24, limit = 10, font_color = "white")}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, February 2024
}
\keyword{players}
\keyword{trending}
