% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_user_waiver_budget.R
\name{plot_user_waiver_budget}
\alias{plot_user_waiver_budget}
\title{Plot User Waiver Budget}
\usage{
plot_user_waiver_budget(
  league_id,
  display_name,
  title = "<b> Waiver Budget</b>",
  tick_color = "black",
  budget_total_line_color = "black",
  fill_color = "lightgreen"
)
}
\arguments{
\item{league_id}{League ID assigned by Sleeper (numeric).}

\item{display_name}{Display name created by user (string).}

\item{title}{Title for plot, which can include HTML formatting (string).}

\item{tick_color}{Font color, name or hex, for display names (string).}

\item{budget_total_line_color}{Line color for max budget, name or hex (string).}

\item{fill_color}{Fill color for bar, name or hex (string).}
}
\value{
Returns a plot containing the waiver budget status.
}
\description{
Given the league ID and display name, plot the current waiver budget
for the user (meaning total waiver budget available versus how much
the user has remaining).
}
\examples{
\dontrun{plot_user_waiver_budget(688281863499907072, "njbultman74")}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, January 2024
}
\keyword{league}
\keyword{rankings}
\keyword{regular}
\keyword{season}
