% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wl_nfp.R
\name{plot_wl_nfp}
\alias{plot_wl_nfp}
\title{Plot League Win/Loss and Net Fantasy Points}
\usage{
plot_wl_nfp(league_id, title = "", subtitle = "")
}
\arguments{
\item{league_id}{League ID generated by Sleeper (integer)}

\item{title}{Optional title of the plot (string)}

\item{subtitle}{Optional subtitle of the plot (string)}
}
\value{
Returns a plot showing net wins and losses colored by fantasy point differentials.
}
\description{
Given a league ID, plot the standings so far for the league. This includes wins compared to losses and 
fantasy point differentials. Note that this plot was created specifically with a NFL fantasy league, so
results may vary when Sleeper decides to incorporate other fantasy sport data.
}
\examples{
plot_wl_nfp(688281863499907072, "Net Wins/Losses (Wins - Losses), Colored by FPTS Differential (Points For - Points Against)")

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{losses}
\keyword{plot}
\keyword{points}
\keyword{standings}
\keyword{wins}
