% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_league.R
\name{get_league}
\alias{get_league}
\title{Gather League Data}
\usage{
get_league(league_id)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (numeric or character)}
}
\value{
Returns a list containing information about the league.
}
\description{
Given a league ID, grab the data for that league. Includes roster positions, season type, if it is in
season, and more.
}
\examples{
\dontrun{get_league(688281863499907072)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, September 2021
}
\keyword{user}
\keyword{users}
