% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loser_playoff_bracket.R
\name{get_loser_playoff_bracket}
\alias{get_loser_playoff_bracket}
\title{Gather Loser Playoff Bracket}
\usage{
get_loser_playoff_bracket(league_id)
}
\arguments{
\item{league_id}{League ID generated by Sleeper (numeric or character)}
}
\value{
Returns a data frame containing playoff information for the loser bracket.
}
\description{
Given a league ID, get the loser playoff bracket for the league.
}
\examples{
\dontrun{get_loser_playoff_bracket(688281863499907072)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2021
}
\keyword{league}
\keyword{loser}
\keyword{playoff}
