% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_drafts.R
\name{get_user_drafts}
\alias{get_user_drafts}
\title{Gather User Draft Information}
\usage{
get_user_drafts(user_id, sport, season)
}
\arguments{
\item{user_id}{User ID generated by Sleeper (numeric)}

\item{sport}{Sport that you would like to query ("nfl" only supported currently) (character)}

\item{season}{Season that you would like to select (numeric)}
}
\value{
Returns a data frame containing information about the draft(s).
}
\description{
Given a user ID, sport abbreviation, and season number, gather draft information for that selection.
}
\examples{
\dontrun{get_user_drafts(688556535013502976, "nfl", 2021)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2021
}
\keyword{draft}
\keyword{user}
