% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_league_records_table.R
\name{plot_league_records_table}
\alias{plot_league_records_table}
\title{Plot League Records Table}
\usage{
plot_league_records_table(league_id)
}
\arguments{
\item{league_id}{League ID generated from Sleeper (numeric).}
}
\value{
Returns a table containing team names and their associated records.
}
\description{
Given the league ID, generate a table showing team
names and their associated records.
}
\examples{
\dontrun{plot_league_records_table(688281863499907072)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, December 2023
}
\keyword{league}
\keyword{records}
\keyword{statistics}
