% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nfl_player_high_school_state.R
\name{plot_nfl_player_high_school_state}
\alias{plot_nfl_player_high_school_state}
\title{Plot NFL Player High School State Information}
\usage{
plot_nfl_player_high_school_state(
  title = "<b>NFL Players by High School State</b>",
  high_fill = "lightgreen",
  low_fill = "white"
)
}
\arguments{
\item{title}{Title for plot (string).}

\item{high_fill}{Fill color, name or hex, for high value states (string).}

\item{low_fill}{Fill color, name or hex, for low value states (string).}
}
\value{
Returns a plot containing information about counts
        for NFL players by their high school state.
}
\description{
This function will generate a plot showing counts of NFL
players by high school state. Colors can be manually defined
for high and low values of each state's count.
}
\examples{
\dontrun{plot_nfl_player_high_school_state()}
\dontrun{plot_nfl_player_high_school_state(title = "Test")}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, March 2024
}
\keyword{high}
\keyword{players}
\keyword{school}
\keyword{state}
