% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nfl_top_colleges.R
\name{plot_nfl_top_colleges}
\alias{plot_nfl_top_colleges}
\title{Plot NFL Player College Information}
\usage{
plot_nfl_top_colleges(
  title = "<b>NFL Players by College</b>",
  tick_color = "black",
  number = 10,
  fill_color = "black"
)
}
\arguments{
\item{title}{Title for plot (string).}

\item{tick_color}{Color for tick labels (string).}

\item{number}{Number of colleges to show (numeric).}

\item{fill_color}{Bar color, number or hex (string).}
}
\value{
Returns a plot containing information about counts
        for NFL players by their college.
}
\description{
This function will generate a bar plot showing counts of NFL
players by college. The total number of colleges returned is
specified by the user.
}
\examples{
\dontrun{plot_nfl_player_college()}
\dontrun{plot_nfl_player_college(title = "TestTitle", tick_color = "green", number = 5)}

}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, March 2024
}
\keyword{college}
\keyword{players}
